# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mural Pay Marketplace API",
    "description": "Backend service for a marketplace that accepts USDC payments on Polygon and automatically converts receipts to Colombian Pesos (COP) via Mural Pay.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://your-app.railway.app", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "MerchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Required for /merchant and /admin endpoints"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "priceUsdc": { "type": "number" },
          "imageUrl": { "type": "string", "nullable": true },
          "stock": { "type": "integer", "description": "-1 means unlimited" },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "OrderSummary": {
        "type": "object",
        "properties": {
          "orderId": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": ["PENDING_PAYMENT", "PAID", "PAYOUT_PENDING", "PAYOUT_COMPLETE", "PAYOUT_FAILED", "EXPIRED"]
          },
          "totalUsdc": { "type": "number" },
          "walletAddress": { "type": "string", "description": "Merchant Polygon wallet to send USDC to" },
          "blockchain": { "type": "string", "example": "POLYGON" },
          "tokenSymbol": { "type": "string", "example": "USDC" },
          "expiresAt": { "type": "string", "format": "date-time" },
          "paymentTxHash": { "type": "string", "nullable": true },
          "customer": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "email": { "type": "string", "format": "email" }
            }
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productName": { "type": "string" },
                "quantity": { "type": "integer" },
                "priceUsdc": { "type": "number" }
              }
            }
          },
          "withdrawal": {
            "nullable": true,
            "type": "object",
            "properties": {
              "status": { "type": "string" },
              "usdcAmount": { "type": "number" },
              "copAmount": { "type": "number", "nullable": true },
              "exchangeRate": { "type": "number", "nullable": true }
            }
          }
        }
      },
      "WithdrawalSummary": {
        "type": "object",
        "properties": {
          "withdrawalId": { "type": "string", "format": "uuid" },
          "orderId": { "type": "string", "format": "uuid" },
          "muralPayoutReqId": { "type": "string" },
          "status": {
            "type": "string",
            "enum": ["INITIATED", "AWAITING_EXECUTION", "PENDING", "COMPLETED", "FAILED", "REFUND_IN_PROGRESS", "REFUNDED"]
          },
          "usdcAmount": { "type": "number" },
          "copAmount": { "type": "number", "nullable": true },
          "exchangeRate": { "type": "number", "nullable": true },
          "customer": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "email": { "type": "string" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "healthCheck",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "status": { "type": "string", "example": "ok" } }
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "List all active products",
        "operationId": "listProducts",
        "tags": ["Catalog"],
        "responses": {
          "200": {
            "description": "Array of products",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a single product",
        "operationId": "getProduct",
        "tags": ["Catalog"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } }
          },
          "404": { "description": "Product not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create an order (customer checkout)",
        "operationId": "createOrder",
        "tags": ["Orders"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer", "items"],
                "properties": {
                  "customer": {
                    "type": "object",
                    "required": ["name", "email"],
                    "properties": {
                      "name": { "type": "string", "example": "Jane Doe" },
                      "email": { "type": "string", "format": "email", "example": "jane@example.com" }
                    }
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity"],
                      "properties": {
                        "productId": { "type": "string", "format": "uuid" },
                        "quantity": { "type": "integer", "minimum": 1 }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created with payment instructions",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/OrderSummary" }],
                  "properties": {
                    "instructions": { "type": "string", "description": "Human-readable payment instructions" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Merchant not configured", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "summary": "Get order status",
        "operationId": "getOrder",
        "tags": ["Orders"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Order detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderSummary" } } } },
          "404": { "description": "Order not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/merchant/orders": {
      "get": {
        "summary": "List all orders (merchant)",
        "operationId": "merchantListOrders",
        "tags": ["Merchant"],
        "security": [{ "MerchantApiKey": [] }],
        "responses": {
          "200": { "description": "Array of orders", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/OrderSummary" } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/merchant/orders/{id}": {
      "get": {
        "summary": "Get order detail (merchant)",
        "operationId": "merchantGetOrder",
        "tags": ["Merchant"],
        "security": [{ "MerchantApiKey": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Order with full detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrderSummary" } } } },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/merchant/withdrawals": {
      "get": {
        "summary": "List all COP withdrawals (merchant)",
        "operationId": "merchantListWithdrawals",
        "tags": ["Merchant"],
        "security": [{ "MerchantApiKey": [] }],
        "responses": {
          "200": { "description": "Array of withdrawals", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WithdrawalSummary" } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/merchant/withdrawals/{id}": {
      "get": {
        "summary": "Get withdrawal detail (merchant)",
        "operationId": "merchantGetWithdrawal",
        "tags": ["Merchant"],
        "security": [{ "MerchantApiKey": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Withdrawal detail", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WithdrawalSummary" } } } },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Withdrawal not found" }
        }
      }
    },
    "/webhooks/mural": {
      "post": {
        "summary": "Mural webhook receiver",
        "operationId": "muralWebhook",
        "tags": ["Webhooks"],
        "description": "Receives MURAL_ACCOUNT_BALANCE_ACTIVITY and PAYOUT_REQUEST events from Mural Pay. Signature verified via ECDSA.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": { "type": "string", "enum": ["MURAL_ACCOUNT_BALANCE_ACTIVITY", "PAYOUT_REQUEST"] },
                  "payoutRequestId": { "type": "string", "description": "Present for PAYOUT_REQUEST events" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Acknowledged" }
        }
      }
    },
    "/admin/products": {
      "post": {
        "summary": "Create a product",
        "operationId": "adminCreateProduct",
        "tags": ["Admin"],
        "security": [{ "MerchantApiKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "priceUsdc"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "priceUsdc": { "type": "number" },
                  "imageUrl": { "type": "string" },
                  "stock": { "type": "integer", "description": "-1 for unlimited" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created product", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Product" } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/admin/setup": {
      "post": {
        "summary": "Initialize/re-run Mural merchant setup",
        "operationId": "adminSetup",
        "tags": ["Admin"],
        "security": [{ "MerchantApiKey": [] }],
        "responses": {
          "200": {
            "description": "Setup result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "steps": { "type": "array", "items": { "type": "string" } },
                    "config": {
                      "type": "object",
                      "properties": {
                        "muralAccountId": { "type": "string" },
                        "walletAddress": { "type": "string" },
                        "counterpartyId": { "type": "string" },
                        "payoutMethodId": { "type": "string" },
                        "webhookId": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/admin/config": {
      "get": {
        "summary": "View current merchant configuration",
        "operationId": "adminGetConfig",
        "tags": ["Admin"],
        "security": [{ "MerchantApiKey": [] }],
        "responses": {
          "200": { "description": "Current config" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not configured" }
        }
      }
    }
  }
}